feat: Generate constructor interfaces
In TypeScript, the parameters to a constructor can be retrieved with a
utility type called `ConstructorParameters<Type>`.
https://www.typescriptlang.org/docs/handbook/utility-types.html#constructorparameterstype

The Type parameter to ConstructorParameters expects a Constructor
interface type, which is an interface with a `new` member and a
`prototype` member.

By defining these *Constructor interfaces, Three.js class types can now
be used with ConstructorParameters.