fix: Suppress double `@remarks` warning on typedoc
typedoc emits: `At most one @remarks tag is expected in a comment, ignoring all but the first in comment`.
It seems we can't have two or more `@remarks` in a doc comment.

context: The error above occurs when I use typedoc on project that refers the Three.js type definitions.